Fixed the Ostats_plot (#FIXME) and added arguments
Removed the argument size because I found no difference in the output with or without it. I also tried a lot of values that did not modified the output. So, I removed it.
The sites2use argument was also giving error because the traits were not filtered according to the sites selected. Now it works.

Please try it out and see if it works out in your computer too. :)