Create testthat_vignette_issue4.R
I created this script to test the remaining parts of the vignette pointed out by Quentin th the issue Write a test suite #4 . As it is the first time I test a code with testthat, I have some doubts I included in the script to discuss.
I also ask for checking if I did and tested it it right.